Move schemas to schemas/base subpackage
Move core schema files (base, info, request, request_stats, response,
statistics, tool_call) into schemas/base/ subdirectory. Move
conversation_graph.py to scheduler/schemas/ where it belongs. Update
all internal and external imports accordingly.

Generated-by: Claude Opus 4.6
