+++ src/guidellm/schemas/base/__init__.py
@@ -1,43 +1,42 @@
 """
-Pydantic schema models for GuideLLM operations.
+Base Pydantic schema models for GuideLLM operations.
 
 Provides standardized data models and type definitions for generation requests,
-responses, timing measurements, and statistics aggregation. These schemas ensure
-type safety and consistent data handling across the benchmarking pipeline,
-from request submission through backend processing to results compilation.
+responses, timing measurements, and statistics aggregation.
 """
 
 from __future__ import annotations
 
-from guidellm.schemas.base import (
+from .base import (
     BaseModelT,
-    DistributionSummary,
     ErroredT,
-    FunctionObjT,
-    GenerationRequest,
-    GenerationRequestArguments,
-    GenerationResponse,
-    GenerativeRequestStats,
     IncompleteT,
-    Percentiles,
     PydanticClassRegistryMixin,
     RegisterClassT,
     ReloadableBaseModel,
-    RequestInfo,
-    RequestSettings,
-    RequestTimings,
     StandardBaseDict,
     StandardBaseModel,
     StatusBreakdown,
-    StatusDistributionSummary,
     SuccessfulT,
-    ToolCall,
-    ToolCallFunction,
     TotalT,
-    TurnType,
-    UsageMetrics,
     standard_model_config,
 )
+from .info import RequestInfo, RequestSettings, RequestTimings
+from .request import (
+    GenerationRequest,
+    GenerationRequestArguments,
+    TurnType,
+    UsageMetrics,
+)
+from .request_stats import GenerativeRequestStats
+from .response import GenerationResponse
+from .statistics import (
+    DistributionSummary,
+    FunctionObjT,
+    Percentiles,
+    StatusDistributionSummary,
+)
+from .tool_call import ToolCall, ToolCallFunction
 
 __all__ = [
     "BaseModelT",
